requirements.txt
coreapi
Django
django-filter
djangorestframework
djangorestframework-jsonp
Markdown
